requirements.txt
Flask==1.0.2
Pillow==5.3.0
requests==2.20.0
docopt==0.6.2
